+++ CHANGELOG.md
@@ -12,4 +12,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SeparatorForIsland [コミット](https://github.com/ReinaS-64892/TTT-DestructiveTextureUtilities/commit/acfcafaa3c150bfdefd89433a3dcfa1a697e925f)
 - StackExtractor [コミット](https://github.com/ReinaS-64892/TTT-DestructiveTextureUtilities/commit/4d2883051aa1ab4628d782c00459da9f4c6c7430)
 - BakeToPNG [コミット](https://github.com/ReinaS-64892/TTT-DestructiveTextureUtilities/commit/c8d65da6e1d84bc81f067ab9964e86cd2525140a)
+- BakeToPNG に InPlaceMode が追加されました (#6)
 - ImportedImageExtractor [コミット](https://github.com/ReinaS-64892/TTT-DestructiveTextureUtilities/commit/cca1f4330ff0fdc229c0f06f9bf29a731a4499e0)